UI changes

+++ src/main/java/com/universityregistration/ui/students/RemoveStudentLayoutFactory.java
@@ -4,11 +4,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import com.universityregistration.model.entity.Student;
 import com.universityregistration.services.RemoveStudentService;
 import com.universityregistration.services.ShowStudentsService;
-import com.universityregistration.ui.commons.UniversMainUI;
+import com.universityregistration.ui.commons.UniversityMainUI;
 import com.universityregistration.utils.Constants;
 import com.vaadin.data.util.BeanItemContainer;
 import com.vaadin.navigator.View;
 import com.vaadin.navigator.ViewChangeListener.ViewChangeEvent;
+import com.vaadin.server.Page;
 import com.vaadin.spring.annotation.SpringView;
 import com.vaadin.ui.Button;
 import com.vaadin.ui.Button.ClickEvent;
@@ -20,43 +21,79 @@ import com.vaadin.ui.Notification.Type;
 import com.vaadin.ui.VerticalLayout;
 import com.vaadin.ui.themes.ValoTheme;
 
-@SpringView(name = RemoveStudentLayoutFactory.NAME, ui = UniversMainUI.class)
+@SpringView(name = RemoveStudentLayoutFactory.NAME, ui = UniversityMainUI.class)
 public class RemoveStudentLayoutFactory extends VerticalLayout implements View, Button.ClickListener {
 	@Autowired
 	private ShowStudentsService studentService;
 	@Autowired
 	private RemoveStudentService removeStudentService;
 	public static final String NAME = "removestudent";
-	private Grid removeStudentTable;
-	private Button removeStudentsButton;
+	private Grid table;
+	private Button removeButton;
+	private Button cancelButton;
 	private List<Student> students;
+	BeanItemContainer<Student> container;
+	
+	private void init(){
+		removeButton = new Button(Constants.REMOVE_STUDENT.getStr());
+		cancelButton = new Button(Constants.REMOVE_STUDENT_CANCEL.getStr());
+		container = new BeanItemContainer<Student>(Student.class, students);
+		table = new Grid(container);
+	}
 	
 	private void addLayout(){
-		removeStudentsButton = new Button(Constants.REMOVE_STUDENT.getString());
 		setMargin(true);
-		BeanItemContainer<Student> container = new BeanItemContainer<Student>(Student.class, students);
-		removeStudentTable = new Grid(container);
-		removeStudentTable.setColumnOrder("firstName", "lastName", "age", "gender");
-		removeStudentTable.removeColumn("id");
-		removeStudentTable.removeColumn("university");
-		removeStudentTable.setImmediate(true);
-		removeStudentTable.setSelectionMode(SelectionMode.MULTI);
-		removeStudentsButton.addClickListener(this);
-		removeStudentsButton.setStyleName(ValoTheme.BUTTON_FRIENDLY);
-		addComponent(removeStudentTable);
-		addComponent(removeStudentsButton);
+		table.setColumnOrder("firstName", "lastName", "age", "gender");
+		table.removeColumn("id");
+		table.removeColumn("university");
+		table.setImmediate(true);
+		table.setSelectionMode(SelectionMode.MULTI);
+		table.setSizeFull();
+		// cancelButton.addClickListener(this);
+		removeButton.addClickListener(this);
+		removeButton.setStyleName(ValoTheme.BUTTON_FRIENDLY);
+		// cancelButton.setStyleName(ValoTheme.BUTTON_FRIENDLY);
+		// GridLayout layout = new GridLayout(4, 4);
+		// layout.setSizeUndefined();
+		// layout.setSpacing(true);
+		addComponent(table);
+		addComponent(removeButton);
+		// layout.addComponent(table, 3, 0);
+		// layout.addComponent(new HorizontalLayout(cancelButton, removeButton), 0, 4);
+		// addComponent(layout);
 	}
 	
 	public void buttonClick(ClickEvent event){
-		MultiSelectionModel selectionModel = (MultiSelectionModel) removeStudentTable.getSelectionModel();
-		for(Object selectedItem: selectionModel.getSelectedRows()){
-			Student student = (Student) selectedItem;
-			removeStudentTable.getContainerDataSource().removeItem(student);
-			removeStudentService.removeStudent(student);
+		if(event.getSource() == this.removeButton){
+			remove();
+		} else{
+			clearFields();
 		}
-		Notification.show(Constants.STUDENT_REMOVE_SUCCESS_TITLE.getString(),
-				Constants.STUDENT_REMOVE_SUCCESS_DESCRIPTION.getString(), Type.WARNING_MESSAGE);
-		removeStudentTable.getSelectionModel().reset();
+	}
+	
+	private void remove(){
+		Notification n;
+		MultiSelectionModel selectionModel = (MultiSelectionModel) table.getSelectionModel();
+		if(selectionModel.getSelectedRows().isEmpty()){
+			n = new Notification(Constants.ERROR.getStr(), Constants.NO_STUDENT_SELECTED.getStr(), Type.ERROR_MESSAGE, true);
+			n.setDelayMsec(200000);
+			n.setStyleName(ValoTheme.NOTIFICATION_ERROR + " " + ValoTheme.NOTIFICATION_CLOSABLE);
+			n.show(Page.getCurrent());
+		} else{
+			for(Object selectedItem: selectionModel.getSelectedRows()){
+				Student student = (Student) selectedItem;
+				table.getContainerDataSource().removeItem(student);
+				removeStudentService.removeStudent(student);
+			}
+			n = new Notification(Constants.STUDENT_SUCCESSFULLY_REMOVED.getStr(), Type.WARNING_MESSAGE);
+			n.setStyleName(ValoTheme.NOTIFICATION_SUCCESS + " " + ValoTheme.NOTIFICATION_CLOSABLE);
+			n.setDelayMsec(200000);
+			n.show(Page.getCurrent());
+		}
+	}
+	
+	private void clearFields(){
+		table.getSelectionModel().reset();
 	}
 	
 	private void loadStudents(){
@@ -64,9 +101,10 @@ public class RemoveStudentLayoutFactory extends VerticalLayout implements View, 
 	}
 	
 	public void enter(ViewChangeEvent event){
-		if(removeStudentTable != null)
+		if(table != null)
 			return;
 		loadStudents();
+		init();
 		addLayout();
 	}
 }
